# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EchoID Nexus API
  description: |
    AI-powered academic assistant for UIT University.
    This document is the **single source of truth** for all API contracts.
    No endpoint is implemented before it appears here.
  version: 0.1.0
  contact:
    name: Team Neurorise
    email: admin@uit.ac.ma
servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://backend:8080
    description: Docker internal
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created — verification email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUserProfileDto'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/verify:
    get:
      tags: [Auth]
      summary: Verify email address
      operationId: verifyEmail
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Email verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuthTokenDto'
        '401':
          description: Invalid credentials or email not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New token pair issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuthTokenDto'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout and invalidate refresh token
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token invalidated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'

  # ── Users ───────────────────────────────────────
...
  /api/courses:
    get:
      tags: [Courses]
      summary: List enrolled courses
      operationId: listCourses
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Course list returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCourseList'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/courses/{id}:
    get:
      tags: [Courses]
      summary: Get course detail with sections
      operationId: getCourseDetail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Course detail returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCourseDetailDto'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/courses/{id}/assignments:
    get:
      tags: [Courses]
      summary: List assignments for a course
      operationId: listAssignments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Assignment list returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAssignmentList'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── LMS ─────────────────────────────────────────
...
  /api/lms/moodle/courses/{id}:
    get:
      tags: [LMS]
      summary: Moodle course detail with section tree
      operationId: getMoodleCourse
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Course detail returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCourseDetailDto'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/lms/gclassroom/courses/{id}:
    get:
      tags: [LMS]
      summary: Google Classroom course detail
      operationId: getGClassroomCourse
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Course detail returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCourseDetailDto'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Events ──────────────────────────────────────
  /api/events:
    post:
      tags: [Events]
      summary: Ingest a behavioral event
      operationId: ingestEvent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
      responses:
        '201':
          description: Event recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          description: Unknown eventType or validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── AI Proxy ────────────────────────────────────
  /api/ai/explain:
    post:
      tags: [AI]
      summary: Get AI concept explanation
      operationId: explainConcept
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExplainRequest'
      responses:
        '200':
          description: Explanation card returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseExplanationCardDto'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/ai/ocr/upload:
    post:
      tags: [AI]
      summary: Upload document for OCR processing
      operationId: uploadForOcr
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, courseId, sectionId]
              properties:
                file:
                  type: string
                  format: binary
                courseId:
                  type: string
                  format: uuid
                sectionId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: OCR result returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOcrResultDto'
        '400':
          description: Unsupported file type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/ai/videos:
    post:
      tags: [AI]
      summary: Search YouTube videos for a concept
      operationId: searchVideos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoSearchRequest'
      responses:
        '200':
          description: Video list returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVideoList'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Workload & Device Token ─────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, displayName, school]
      properties:
        email:
          type: string
          format: email
          example: student@ensa.uit.ac.ma
        password:
          type: string
          minLength: 8
          maxLength: 128
          example: demo1234
        displayName:
          type: string
          example: Ahmed Benali
        school:
          type: string
          enum: [ENSA, EST, FAC, OTHER]
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: student@ensa.uit.ac.ma
        password:
          type: string
          example: demo1234
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
    EventRequest:
      type: object
      required: [eventType, timestamp]
      properties:
        eventType:
          type: string
          enum: [concept_query, professor_upload, notification_sent, session_event]
        studentId:
          type: string
          format: uuid
        professorId:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
        sectionId:
          type: string
          format: uuid
        conceptText:
          type: string
          example: Big-O complexity
        explanationLevel:
          type: string
          enum: [BEGINNER, VISUAL, ADVANCED]
        fileType:
          type: string
          example: PDF
        triggerType:
          type: string
          example: WORKLOAD_CRITICAL
        payloadSummary:
          type: string
        sessionEventType:
          type: string
          enum: [LOGIN, LOGOUT, SCREEN_VIEW]
        timestamp:
          type: string
          format: date-time
    ExplainRequest:
      type: object
      required: [conceptText, courseId, sectionId, level]
      properties:
        conceptText:
          type: string
          example: Big-O notation
        courseId:
          type: string
          format: uuid
        sectionId:
          type: string
          format: uuid
        level:
          type: string
          enum: [BEGINNER, VISUAL, ADVANCED]
    VideoSearchRequest:
      type: object
      required: [conceptText, courseId]
      properties:
        conceptText:
          type: string
          example: Big-O notation
        courseId:
          type: string
          format: uuid
...
    AuthTokenDto:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIs...
        refreshToken:
          type: string
          format: uuid
        expiresIn:
          type: integer
          format: int64
          example: 86400000
        tokenType:
          type: string
          example: Bearer
    UserProfileDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          example: student@ensa.uit.ac.ma
        fullName:
          type: string
          example: Ahmed Benali
        picture:
          type: string
          nullable: true
        role:
          type: string
          enum: [STUDENT, PROFESSOR, ADMIN]
        school:
          type: string
          enum: [ENSA, EST, FAC, OTHER]
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CourseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Introduction to Machine Learning
        lmsSource:
          type: string
          enum: [MOODLE, GOOGLE_CLASSROOM]
        school:
          type: string
          enum: [ENSA, EST, FAC, OTHER]
        semester:
          type: string
          enum: [S1, S2, ANNUAL]
        isActive:
          type: boolean
    CourseDetailDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        lmsSource:
          type: string
          enum: [MOODLE, GOOGLE_CLASSROOM]
        school:
          type: string
        sections:
          type: array
          items:
            $ref: '#/components/schemas/CourseSectionDto'
    CourseSectionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: "Chapter 3: Sorting Algorithms"
        orderIndex:
          type: integer
    AssignmentDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
        courseTitle:
          type: string
        title:
          type: string
          example: Linear Regression Report
        description:
          type: string
          nullable: true
        dueAt:
          type: string
          format: date-time
        complexity:
          type: number
          format: double
          example: 3.2
        assignmentType:
          type: string
          enum: [EXAM, PROJECT, HOMEWORK]
    ExplanationCardDto:
      type: object
      properties:
        title:
          type: string
          example: Understanding Big-O Notation
        body:
          type: string
        level:
          type: string
          enum: [BEGINNER, VISUAL, ADVANCED]
        bulletPoints:
          type: array
          items:
            type: string
    VideoResultDto:
      type: object
      properties:
        videoId:
          type: string
          example: dQw4w9WgXcQ
        title:
          type: string
        channelName:
          type: string
        thumbnailUrl:
          type: string
        durationSeconds:
          type: integer
    OcrResultDto:
      type: object
      properties:
        documentId:
          type: string
          format: uuid
          nullable: true
        extractedText:
          type: string
        pageCount:
          type: integer
        confidence:
          type: number
          format: double
...
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        data:
          nullable: true
        message:
          type: string
    ApiResponseUserProfileDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/UserProfileDto'
        message:
          type: string
    ApiResponseAuthTokenDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/AuthTokenDto'
        message:
          type: string
    ApiResponseCourseList:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/CourseDto'
        message:
          type: string
    ApiResponseCourseDetailDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/CourseDetailDto'
        message:
          type: string
    ApiResponseAssignmentList:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/AssignmentDto'
        message:
          type: string
    ApiResponseExplanationCardDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/ExplanationCardDto'
        message:
          type: string
    ApiResponseVideoList:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/VideoResultDto'
        message:
          type: string
    ApiResponseOcrResultDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/OcrResultDto'
        message:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
        path:
          type: string